Improvements to the documentaton for core.io package (#750)
