+++ src/main/java/net/openhft/chronicle/core/io/CleaningRandomAccessFile.java
@@ -22,9 +22,13 @@ import java.io.FileNotFoundException;
 import java.io.RandomAccessFile;
 
 /**
- * A RandomAccessFile must be explicitly close or cause a resources leak.
+ * {@link RandomAccessFile} that quietly closes itself during finalisation if it
+ * has not already been closed. This guards against leaks when a test forgets to
+ * close the file, although callers should still use try-with-resources wherever
+ * possible.
  * <p>
- * Weak references RAF can result in a resource leak when GC'ed which doesn't appear if the GC isn't running.
+ * Finalisation is deprecated in recent JDKs so this class is only a safety net
+ * for legacy code.
  */
 public class CleaningRandomAccessFile extends RandomAccessFile {
     public CleaningRandomAccessFile(String name, String mode) throws FileNotFoundException {
@@ -38,6 +42,7 @@ public class CleaningRandomAccessFile extends RandomAccessFile {
     @SuppressWarnings({"deprecation", "removal"})
     @Override
     protected void finalize() throws Throwable {
+        // best-efforts attempt to close the file if the owner forgot
         super.finalize();
         Closeable.closeQuietly(this);
     }
