+++ src/main/java/net/openhft/chronicle/core/io/ClosedState.java
@@ -17,8 +17,8 @@
 
 package net.openhft.chronicle.core.io;
 /**
- * Functional interface representing a closed state.
- * This interface provides a method to check if an object is in a closed state.
+ * Functional interface used to query whether a resource has been closed.  It is
+ * typically implemented with a lambda or method reference.
  */
 @FunctionalInterface
 public interface ClosedState {
