+++ src/main/java/net/openhft/chronicle/core/io/Monitorable.java
@@ -17,10 +17,9 @@
 package net.openhft.chronicle.core.io;
 
 /**
- * Interface for objects that can be monitored and unmonitored.
- * <p>
- * This is useful for managing resources that need to be tracked and potentially cleaned up
- * when they are no longer needed.
+ * Implement for resources that may be tracked by tooling such as
+ * {@code CloseableUtils}. Calling {@link #unmonitor()} stops that tracking and
+ * is typically used just before a resource is discarded.
  */
 public interface Monitorable {
