+++ src/main/java/net/openhft/chronicle/core/io/SimpleCloseable.java
@@ -16,8 +16,8 @@
 package net.openhft.chronicle.core.io;
 
 /**
- * An abstract class providing a basic implementation of the {@link Closeable}, {@link ReferenceOwner}, and {@link ManagedCloseable} interfaces.
- * It maintains a flag to track the closed state of the resource and provides methods for closing the resource.
+ * Lightweight helper implementing {@link Closeable} and {@link ManagedCloseable}.
+ * A boolean tracks the closed state. Override {@link #performClose()} for custom clean-up.
  */
 public abstract class SimpleCloseable implements Closeable, ReferenceOwner, ManagedCloseable {
     private transient volatile boolean closed;
@@ -30,8 +30,7 @@ public abstract class SimpleCloseable implements Closeable, ReferenceOwner, Mana
     }
 
     /**
-     * Closes the resource so that it cannot be used again.
-     * Once closed, subsequent calls to this method will have no effect.
+     * Idempotent close for use with try-with-resources.
      */
     @Override
     public final void close() {
@@ -42,11 +41,7 @@ public abstract class SimpleCloseable implements Closeable, ReferenceOwner, Mana
     }
 
     /**
-     * Performs the actual close operation.
-     * Subclasses can override this method to provide custom close logic.
-     * The default implementation does nothing.
-     * <p>
-     * Call close() to ensure this is called exactly once.
+     * Hook for subclasses to release resources. Called once from {@link #close()}.
      */
     protected void performClose() {
         // might be nothing.
